+++ src/fobi/contrib/plugins/form_elements/content/content_video/fobi_form_elements.py
@@ -10,8 +10,9 @@ from django.utils.translation import ugettext_lazy as _
 
 from vishap import render_video
 
+from nonefield.fields import NoneField
+
 from fobi.base import FormElementPlugin, form_element_plugin_registry
-from fobi.fields import NoneField
 from fobi.contrib.plugins.form_elements.content.content_video import UID
 from fobi.contrib.plugins.form_elements.content.content_video.forms \
     import ContentVideoForm