+++ src/fobi/widgets.py
@@ -2,11 +2,10 @@ __title__ = 'fobi.widgets'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 __copyright__ = 'Copyright (c) 2014-2015 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
-__all__ = ('NoneWidget', 'NumberInput', 'BooleanRadioSelect',)
+__all__ = ('NumberInput', 'BooleanRadioSelect',)
 
-from django.forms.widgets import Widget, RadioSelect
+from django.forms.widgets import RadioSelect
 from django.utils.translation import ugettext_lazy as _
-from django.utils.safestring import mark_safe
 
 # Safe import of ``NumberInput``
 try:
@@ -15,15 +14,6 @@ except ImportError:
     from django.forms.widgets import TextInput
     class NumberInput(TextInput):
         input_type = 'number'
-
-class NoneWidget(Widget):
-    """
-    To be used with content elements.
-    """
-    def render(self, name, value, attrs=None):
-        """
-        """
-        return mark_safe(value)
 
 
 BOOLEAN_CHOICES = (